backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
celery[redis]==5.4.0
redis==5.1.1
httpx==0.27.2
tenacity==9.0.0
google-generativeai==0.8.3
beautifulsoup4==4.12.3
python-whois==0.9.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
